+++ Source/UI/CyclingValueButton.cs
@@ -6,7 +6,7 @@ public class CyclingValueButton<T> : Button
     private int _index;
     private readonly List<T> _cycleValues;
 
-    public CyclingValueButton(int layer, Rect position, UITheme theme, List<T> cycleValues)
+    public CyclingValueButton(float layer, Rect position, UITheme theme, List<T> cycleValues)
         : base(layer, position, theme.Button.DefaultColour, theme, "")
     {
         if (cycleValues.Count == 0)
