+++ Source/UI/ValueDisplay/ValueDisplayWithButtons.cs
@@ -4,7 +4,7 @@ public delegate void ValueChange(int change);
 
 public class ValueDisplayWithButtons : ValueDisplay
 {
-    public ValueDisplayWithButtons(int layer, Rect position, string graphic, UITheme theme, string valueName, ValueGet valueToTrack, Rect plusButtonPos, string plusButtonGfx, Rect minusButtonPos, string minusButtonGfx, ValueChange changeFunction)
+    public ValueDisplayWithButtons(float layer, Rect position, string graphic, UITheme theme, string valueName, ValueGet valueToTrack, Rect plusButtonPos, string plusButtonGfx, Rect minusButtonPos, string minusButtonGfx, ValueChange changeFunction)
         : base(layer, position, graphic, theme, valueName, valueToTrack)
     {
         Entity plusButton = new Button(layer, plusButtonPos, plusButtonGfx, () => { changeFunction(1); UpdateValueText(); });
